+++ src/reports/htmlbuilder.h
@@ -108,6 +108,7 @@ public:
     void endTableCell();
 
     const wxString getHTMLText() const;
+    const wxString getBugReportTemplate() const;
 
     void addTableRow(const wxString& label, double data);
